+++ satellite/src/main/java/com/linbit/linstor/layer/storage/ebs/EbsInitiatorProvider.java
@@ -595,15 +595,16 @@ public class EbsInitiatorProvider extends AbsEbsProvider<LsBlkEntry>
     }
 
     @Override
-    public void update(StorPool storPoolRef) throws AccessDeniedException, DatabaseException, StorageException
+    public @Nullable LocalPropsChangePojo update(StorPool storPoolRef)
+        throws AccessDeniedException, DatabaseException, StorageException
     {
-        // noop
+        return null;
     }
 
     @Override
-    public LocalPropsChangePojo checkConfig(StorPool storPoolRef) throws StorageException, AccessDeniedException
+    public @Nullable LocalPropsChangePojo checkConfig(StorPool storPoolRef)
+        throws StorageException, AccessDeniedException
     {
-        // TODO Auto-generated method stub
         return null;
     }