+++ satellite/src/main/java/com/linbit/linstor/layer/storage/exos/ExosProvider.java
@@ -61,6 +61,7 @@ import com.linbit.linstor.transaction.manager.TransactionMgr;
 import com.linbit.utils.Align;
 import com.linbit.utils.Triple;
 
+import javax.annotation.Nullable;
 import javax.inject.Inject;
 import javax.inject.Provider;
 import javax.inject.Singleton;
@@ -907,7 +908,7 @@ public class ExosProvider extends AbsStorageProvider<ExosRestVolume, ExosData<Re
         return reinit();
     }
 
-    private LocalPropsChangePojo reinit() throws StorageException, AccessDeniedException
+    private @Nullable LocalPropsChangePojo reinit() throws StorageException, AccessDeniedException
     {
         LocalPropsChangePojo ret = null;
         if (extToolsChecker.areSupported(false, ExtTools.LSSCSI, ExtTools.SAS_PHY, ExtTools.SAS_DEVICE))
@@ -1100,7 +1101,7 @@ public class ExosProvider extends AbsStorageProvider<ExosRestVolume, ExosData<Re
     }
 
     @Override
-    public LocalPropsChangePojo checkConfig(StorPool storPool) throws StorageException, AccessDeniedException
+    public @Nullable LocalPropsChangePojo checkConfig(StorPool storPool) throws StorageException, AccessDeniedException
     {
         Props props = DeviceLayerUtils.getNamespaceStorDriver(
             storPool.getProps(storDriverAccCtx)
@@ -1161,9 +1162,10 @@ public class ExosProvider extends AbsStorageProvider<ExosRestVolume, ExosData<Re
     }
 
     @Override
-    public void update(StorPool storPoolRef) throws AccessDeniedException, DatabaseException, StorageException
+    public @Nullable LocalPropsChangePojo update(StorPool storPoolRef)
+        throws AccessDeniedException, DatabaseException, StorageException
     {
-        // noop
+        return null;
     }
 
     @Override