+++ satellite/src/main/java/com/linbit/linstor/layer/storage/spdk/AbsSpdkProvider.java
@@ -48,6 +48,7 @@ import com.linbit.linstor.transaction.manager.TransactionMgr;
 
 import static com.linbit.linstor.layer.storage.spdk.utils.SpdkUtils.SPDK_PATH_PREFIX;
 
+import javax.annotation.Nullable;
 import javax.inject.Provider;
 
 import java.io.File;
@@ -428,7 +429,15 @@ public abstract class AbsSpdkProvider<T> extends AbsStorageProvider<LvsInfo, Spd
     }
 
     @Override
-    public LocalPropsChangePojo checkConfig(StorPool storPool) throws StorageException, AccessDeniedException
+    public @Nullable LocalPropsChangePojo update(StorPool storPoolRef)
+        throws AccessDeniedException, DatabaseException, StorageException
+    {
+        // TODO: we need to implement a check for pmem here. something like LvmProvider.update does
+        return null;
+    }
+
+    @Override
+    public @Nullable LocalPropsChangePojo checkConfig(StorPool storPool) throws StorageException, AccessDeniedException
     {
         LocalPropsChangePojo ret = new LocalPropsChangePojo();
 
@@ -440,7 +449,7 @@ public abstract class AbsSpdkProvider<T> extends AbsStorageProvider<LvsInfo, Spd
 
         checkExtentSize(storPool, ret);
 
-        return null;
+        return ret;
     }
 
     protected void checkExtentSize(StorPool storPool, LocalPropsChangePojo ret)
@@ -495,12 +504,6 @@ public abstract class AbsSpdkProvider<T> extends AbsStorageProvider<LvsInfo, Spd
             }
         }
         return volumeGroups;
-    }
-
-    @Override
-    public void update(StorPool storPoolRef) throws AccessDeniedException, DatabaseException, StorageException
-    {
-        // TODO: we need to implement a check for pmem here. something like LvmProvider.update does
     }
 
     /*