+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+
 - Freespace calc: Snapshots only reserve allocated size now instead of volume size
 - resource-definition modify now, only warns if a satellite is offline
 - AutoVerifyAlgo: allow to use algorithm name additionally to the specific implementation
+- ZFS: try to determine default volblocksize
 
 ### Fixed
+
 - Deleting a remote can now be safely done even if there are active shippings to said remote
 - MakeAvail: trigger update satellite if only FLAGS changed
 - Resource delete: Don't check user set verify-algo if last resource deleted.